Extract `ngff` from `ome_zarr` subpackage
