+++ fractal_tasks_core/tasks/apply_registration_to_ROI_tables.py
@@ -24,7 +24,7 @@ import pandas as pd
 import zarr
 from pydantic.decorator import validate_arguments
 
-from fractal_tasks_core.ome_zarr.ngff import load_NgffWellMeta
+from fractal_tasks_core.ngff import load_NgffWellMeta
 from fractal_tasks_core.roi import (
     are_ROI_table_columns_valid,
 )
