+++ fractal_tasks_core/tasks/apply_registration_to_image.py
@@ -26,7 +26,7 @@ import numpy as np
 import zarr
 from pydantic.decorator import validate_arguments
 
-from fractal_tasks_core.ome_zarr.ngff import load_NgffImageMeta
+from fractal_tasks_core.ngff import load_NgffImageMeta
 from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
 from fractal_tasks_core.ome_zarr.zattrs_utils import get_table_path_dict
 from fractal_tasks_core.roi import (
