+++ fractal_tasks_core/tasks/calculate_registration_image_based.py
@@ -25,9 +25,9 @@ import zarr
 from pydantic.decorator import validate_arguments
 from skimage.registration import phase_cross_correlation
 
+from fractal_tasks_core.ngff import load_NgffImageMeta
 from fractal_tasks_core.ome_zarr.channels import get_channel_from_image_zarr
 from fractal_tasks_core.ome_zarr.channels import OmeroChannel
-from fractal_tasks_core.ome_zarr.ngff import load_NgffImageMeta
 from fractal_tasks_core.roi import check_valid_ROI_indices
 from fractal_tasks_core.roi import (
     convert_indices_to_regions,
