+++ fractal_tasks_core/tasks/copy_ome_zarr.py
@@ -22,7 +22,7 @@ import zarr
 from pydantic.decorator import validate_arguments
 
 import fractal_tasks_core
-from fractal_tasks_core.ome_zarr.ngff import load_NgffImageMeta
+from fractal_tasks_core.ngff import load_NgffImageMeta
 from fractal_tasks_core.ome_zarr.write import open_zarr_group_with_overwrite
 from fractal_tasks_core.roi import (
     convert_ROIs_from_3D_to_2D,
