+++ tests/tasks/_validation.py
@@ -12,7 +12,7 @@ from devtools import debug
 from jsonschema import validate
 
 from fractal_tasks_core import __OME_NGFF_VERSION__
-from fractal_tasks_core.ome_zarr.ngff import load_NgffImageMeta
+from fractal_tasks_core.ngff.zarr_utils import load_NgffImageMeta
 from fractal_tasks_core.roi import (
     convert_ROI_table_to_indices,
 )
