+++ tests/tasks/test_unit_illumination_correction.py
@@ -10,7 +10,7 @@ import pytest
 from pytest import LogCaptureFixture
 from pytest import MonkeyPatch
 
-from fractal_tasks_core.ome_zarr.ngff import load_NgffImageMeta
+from fractal_tasks_core.ngff.zarr_utils import load_NgffImageMeta
 from fractal_tasks_core.roi import (
     convert_ROI_table_to_indices,
 )
